Add error checking and more format enforcement
Added error checking to ensure that the user is in the correct working directory. This is a requirement when creating the course folder structure.

Added an enforcement on the course name. Specifically, if the user enters a course name that contains whitespaces, the course name is stripped of excess whitespace, and remaining whitespace is replaced with dashes.