Add base image checker
Added functionality to map from args.baseimage to the newest version of the actual base image that will be used for any given course.

Also added a prompt asking the user if the program should delete all the course folders in case the program exits on an error. This is to minimize the risk of accidental folder deletion.