Remove ARG TAG from Dockerfile templates
The tag is substituted into the Dockerfile templates from create-new-course.py, so ARG TAG is not used. Furthermore, the TAG is substituted with the actual tag, so `ARG TAG` will not be present in the templates that are copied into the course folders because `TAG` will be replaced with the actual tag.