Scripts for automatized course setup (#6)
* Create create-new-course.py

* add templates

* Add required format for -r

* Add course start date format check

The user has to give the start date of the course (the -r option) in the format: YYYY-MM-DD.

* Small change

* Add elaborate error checking for -release

* Small change

* Fix small error

* Add code annotation

* Fix spacing

* Add a little string formatting

Cosmetic changes only.

* Fix error in outline of folder structure

* Prepare for editing of template files

The contents of the template files (in /templates) have to be edited based on information from the user such as course name, base image etc. This is done with a simple read-edit-write procedure.

TODO: Implement the 'edit' step.

* Add script documentation

* Add error checking and more format enforcement

Added error checking to ensure that the user is in the correct working directory. This is a requirement when creating the course folder structure.

Added an enforcement on the course name. Specifically, if the user enters a course name that contains whitespaces, the course name is stripped of excess whitespace, and remaining whitespace is replaced with dashes.

* Remove print statement

* Add to error messages

* Add cleanup in case of error

If an error happens occurs after the course's folder structure has been created, the course folder is deleted recursively starting from course_root_dir.

* Minor addition

* Add more format enforcement

Remove all leading and trailing whitespaces from args.name. This will ensure that a course does not start or end with a dash due to whitespace substitution.

* Add base image checker

Added functionality to map from args.baseimage to the newest version of the actual base image that will be used for any given course.

Also added a prompt asking the user if the program should delete all the course folders in case the program exits on an error. This is to minimize the risk of accidental folder deletion.

* Finalize get_baseimage_name()

Also delete course that was accidentally pushed in previous commit.

* Remove tag suffix for VDEs

The '-xfce' suffix is hardcoded into the Dockerfiles instead.

* Edit template files before writing to course folder

* Remove cleanup on Error run

With this cleanup, we risked that teachers accidentally deleted other teacher's courses if they made a course with the same name as an existing course.

* Add error check on course name availability

* Add text substitution in README template

* Remove ARG TAG from Dockerfile templates

The tag is substituted into the Dockerfile templates from create-new-course.py, so ARG TAG is not used. Furthermore, the TAG is substituted with the actual tag, so `ARG TAG` will not be present in the templates that are copied into the course folders because `TAG` will be replaced with the actual tag.

* Create requirements.txt

The remaining modules that are needed to execute create-new-course.py is parts of the Python standard library.

* Add Python version as comment

Python 3.10.14.

* Remove __version__

Because not really meaningful here.

* Remove __author__

Not really meaningful here.

* Minor additions

* change templ

* change script n

* Change folders

- Moved README.md to the course root directory.
- Created /test and /logo is the course root directory
- Change 'startapp' to 'startcourse'

* Remove __status__

Not really meaningful.

* Enforce that <course name> must be lowercase

* Remove unused function

* update templates

* Add directory

* Incorporate suggestions from code review

* Add build-docker-image.py

Script for building docker images from course Dockerfiles. Should be portable across operating systems but currently only tested on macOS. Also update requirements.txt.

* Fix typo

* Add info about Docker daemon

Specifically, the script can only run if a Docker daemon is running.

* Add maintainer info

* Update .gitignore

Ignore .DS_Store everywhere in repo.

* Add uni and course code info

Courses/ now contains subfolders for each university (to being with only AAU, AU, and SDU). Courses for a given university are then placed in the given university folder. Also added the course code (from university course description) to the course folder name (exact naming convention to be decided). University and course code are added as command line argument (-u and -c, respectively).

* Add subfolder for each university

Only AAU, AU, and SDU to begin with.

* Update build-docker-image.py

Add args -u and -c to be able to match course as created using create-new-course.py.
Also changed the image tag.

* Update explanatory comment

Add explanation of -c and -u args.

* Add remaining universities + other

* Ensure uniformity in naming

* Update scripts/build-docker-image.py

Co-authored-by: Jes Elgin Jensen <146733701+jeselginAU@users.noreply.github.com>

* Update scripts/create-new-course.py

Co-authored-by: Jes Elgin Jensen <146733701+jeselginAU@users.noreply.github.com>

* Remove comment

* Incorporate comments from code reveiw

* Add build-and-run-course_template and integrate with create-new-course.py

* removes build-and-run-course_template; adds run-docker-container.py

* adds startCommand required for run-docker-container.py

* script comment fix

* various comment and variable changes

* Using Docker from course folders

Started implemented a more user-friendly solution. Templates are copied into the course folders. These have bare minimum requirements for the teacher to fill out. These call the main scripts for building images and running containers located in `scripts/`.

* Temporary modifications

Will need some final modifications in the end.

* Continue working on scripts/templates

* Continue

* Finalize scripts

* Extra spacing

* More string parsing

* Minor things

* Remove possibility of `build-args`

* Remove excess spacing

* Fix comment alignment

* Rename files

From *`.py` to `*.template`

* Update yaml templates

* Change scripts for building/running

* Update create-new-course.py

* Update docker-build.template

* Add maintainer info

* Minor changes to templates

* Rename 'Courses/Other' -> 'Courses/OTHER'

Makes parsing of paths between course folders and scripts folder easier.

* Add possibility of start command

The teacher can now specify a start command for a Docker container in `docker-run.py` (default: `start_course`).

* Add missing library

* Add more parsing to volumes list

* Minor changes

* Update README.md.template

Use links to UCloud documentation.

* Minor

* Remove file

---------

Co-authored-by: FedericaLoVerso <federica@imada.sdu.dk>
Co-authored-by: Jes Elgin Jensen <146733701+jeselginAU@users.noreply.github.com>
Co-authored-by: Rytis <rm@cas.au.dk>